+++ include/geos/geom/MultiPoint.h
@@ -112,7 +112,7 @@ protected:
 
 	MultiPoint(const MultiPoint &mp): Geometry(mp), GeometryCollection(mp) {}
 
-	const Coordinate* getCoordinateN(int n) const;
+	const Coordinate* getCoordinateN(size_t n) const;
 };
 
 #ifdef _MSC_VER
