+++ include/geos/noding/SegmentNode.h
@@ -61,7 +61,7 @@ public:
 	geom::Coordinate coord;
 
 	/// the index of the containing line segment in the parent edge
-	unsigned int segmentIndex;
+	size_t segmentIndex;
 
 	/// Construct a node on the given NodedSegmentString
 	//
@@ -77,7 +77,7 @@ public:
 	///
 	SegmentNode(const NodedSegmentString& ss,
 		    const geom::Coordinate& nCoord,
-			unsigned int nSegmentIndex, int nSegmentOctant);
+			size_t nSegmentIndex, int nSegmentOctant);
 
 	~SegmentNode() {}
